+++ crates/ruma-events/CHANGELOG.md
@@ -23,6 +23,10 @@ Breaking changes:
 - `RedactContent::redact()` and `FullStateEventContent::redact()` take a
   `RedactionRules` instead of `RoomVersionId`. This avoids undefined behavior
   for unknown room versions.
+- Add implementation of [MSC 4278](https://github.com/matrix-org/matrix-spec-proposals/pull/4278)
+  using `MediaPreviewConfigEventContent` for the stable version and `MediaPreviewConfigEventContent`
+  for the unstable one, both are only implemented as `GlobalAccountData` events as of right now.
+   
 
 Improvements: